Remove `include/grpcpp/impl/codegen/status_code_enum.h` (#31284)
