+++ src/cpp/ext/proto_server_reflection.cc
@@ -23,6 +23,7 @@
 
 #include <grpcpp/grpcpp.h>
 #include <grpcpp/support/interceptor.h>
+#include <grpcpp/support/status_code_enum.h>
 
 // IWYU pragma: no_include <google/protobuf/descriptor.h>
