+++ src/cpp/server/authorization_policy_provider.cc
@@ -21,6 +21,7 @@
 #include <grpcpp/security/authorization_policy_provider.h>
 #include <grpcpp/support/config.h>
 #include <grpcpp/support/status.h>
+#include <grpcpp/support/status_code_enum.h>
 
 namespace grpc {
 namespace experimental {
