+++ src/cpp/server/channelz/channelz_service.cc
@@ -22,6 +22,7 @@
 
 #include <grpc/support/alloc.h>
 #include <grpcpp/impl/codegen/config_protobuf.h>
+#include <grpcpp/support/status_code_enum.h>
 
 // IWYU pragma: no_include <google/protobuf/util/json_util.h>
