+++ src/cpp/server/health/default_health_check_service.cc
@@ -32,6 +32,7 @@
 #include <grpcpp/impl/rpc_method.h>
 #include <grpcpp/impl/rpc_service_method.h>
 #include <grpcpp/support/slice.h>
+#include <grpcpp/support/status_code_enum.h>
 
 #include "src/proto/grpc/health/v1/health.upb.h"
