+++ src/cpp/util/byte_buffer_cc.cc
@@ -27,6 +27,7 @@
 #include <grpcpp/support/byte_buffer.h>
 #include <grpcpp/support/slice.h>
 #include <grpcpp/support/status.h>
+#include <grpcpp/support/status_code_enum.h>
 
 namespace grpc {
